04. Wild Farm

+++ OOP/Polymorphism/04.WildFarm/Models/Animal.cs
@@ -7,7 +7,7 @@ namespace _04.WildFarm.Models
 {
     public abstract class Animal : IAnimal
     {
-        public Animal(string name, double weight)
+        public Animal(string name, decimal weight)
         {
             Name = name;
             Weight = weight;
@@ -15,7 +15,7 @@ namespace _04.WildFarm.Models
         }
 
         public string Name { get; protected set; }
-        public double Weight { get; protected set; }
+        public decimal Weight { get; protected set; }
         public int FoodEaten { get; protected set; }
 
         public abstract void Eat(string foodType, int quantity);
