+++ OOP/Polymorphism/04.WildFarm/Models/Mammals/Dog.cs
@@ -6,10 +6,10 @@ namespace _04.WildFarm.Models
 {
     public class Dog : Mammal
     {
-        private const double WeightIncreaseFactor = 0.40f;
+        private const decimal WeightIncreaseFactor = 0.40M;
         private IList<string> AcceptedFood = new string[] { "Meat" };
 
-        public Dog(string name, double weight, string livingRegion) : base(name, weight, livingRegion)
+        public Dog(string name, decimal weight, string livingRegion) : base(name, weight, livingRegion)
         {
         }
